+++ AzNetworkDiagram.psm1
@@ -4998,6 +4998,124 @@ function Export-MgmtGroups
     }
 }
 
+<#
+.SYNOPSIS
+Exports details of Public IP for inclusion in an IP Plan.
+
+.DESCRIPTION
+The `Export-PIP2IPPlan` function processes a specified PIP object, retrieves its details, and formats the data for inclusion in an IP Plan.
+
+.PARAMETER VNet
+Specifies the Azure PIP object to be processed. This parameter is mandatory.
+
+.EXAMPLE
+PS> $PIPs = Get-AzPublicIpAddress
+PS> $PIPs | Foreach-Object { Export-PIP2IPPlan -PIP $_ }
+
+This example retrieves specified Azure PIP object, retrieves its details, and formats the data for inclusion in an in an IP Plan.
+#>
+function Export-PIP2IPPlan {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory = $true)]
+        [PSCustomObject]$PIP
+    )
+
+    try {
+        #$data = ""
+        #$id = $MgmtGroupEntityObject.Id.replace("-", "").replace("/", "").replace(".", "").ToLower()
+        #$subid = $context.Subscription.Id
+        $subname = (Get-AzContext).Subscription.Name
+                       
+    
+        $obj = [pscustomobject]@{
+            AddressSpace     = ''
+            Type             = ''
+            Resource         = ''
+            ResourceGroup    = ''
+            SubscriptionName = ''
+            SubscriptionId   = ''
+            #vnetid          = ''
+            PrefixName       = ''
+        }
+        $prefix = $PIP.IpAddress
+        $IPPREARMID = $PIP.PublicIpPrefix.id
+        $IPPREName = ""
+        if ( $null -ne $IPPREARMID ) { $IPPREName = $IPPREARMID.split("/")[8] }
+
+        $obj.SubscriptionName = $subname
+        # $obj.SubscriptionId = $subid
+        $obj.Type = "PIP"
+        $obj.ResourceGroup = $PIP.ResourceGroupName
+        $obj.Resource = $PIP.Name
+        $obj.AddressSpace = $prefix
+        $obj.PrefixName = $IPPREName
+        #$obj.vnetid = $vnet.id
+
+        $script:IPPlan += $obj
+        
+    }
+    catch {
+        Write-Error "Can't export Public IP 2 IP Plan at line $($_.InvocationInfo.ScriptLineNumber) " $_.Exception.Message
+    }
+} 
+
+<#
+.SYNOPSIS
+Exports details of Public IP Prefix for inclusion in an IP Plan.
+
+.DESCRIPTION
+The `Export-IPPRE2IPPlan` function processes a specified PIP object, retrieves its details, and formats the data for inclusion in an IP Plan.
+
+.PARAMETER VNet
+Specifies the Azure PIP object to be processed. This parameter is mandatory.
+
+.EXAMPLE
+PS> $IPPREs = Get-AzPublicIPPrefix
+PS> $IPPREs | Foreach-Object { Export-PIP2IPPlan -IPPRE $_ }
+
+This example retrieves specified Azure Public IP Prefix object, retrieves its details, and formats the data for inclusion in an in an IP Plan.
+#>
+function Export-IPPRE2IPPlan {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory = $true)]
+        [PSCustomObject]$IPPRE
+    )
+
+    try {
+        #$data = ""
+        #$id = $MgmtGroupEntityObject.Id.replace("-", "").replace("/", "").replace(".", "").ToLower()
+        #$subid = $context.Subscription.Id
+        $subname = (Get-AzContext).Subscription.Name
+                       
+    
+        $obj = [pscustomobject]@{
+            AddressSpace     = ''
+            Type             = ''
+            Resource         = ''
+            ResourceGroup    = ''
+            SubscriptionName = ''
+            SubscriptionId   = ''
+            #vnetid          = ''
+        }
+        $prefix = $IPPRE.IPPrefix
+
+        $obj.SubscriptionName = $subname
+        # $obj.SubscriptionId = $subid
+        $obj.Type = "IPPRE"
+        $obj.ResourceGroup = $IPPRE.ResourceGroupName
+        $obj.Resource = $IPPRE.Name
+        $obj.AddressSpace = $prefix
+        #$obj.vnetid = $vnet.id
+
+        $script:IPPlan += $obj
+        
+    }
+    catch {
+        Write-Error "Can't export Public IP Prefix at line $($_.InvocationInfo.ScriptLineNumber) " $_.Exception.Message
+    }
+} 
 
 <#
 .SYNOPSIS
@@ -5217,6 +5335,8 @@ function Export-IPPlan {
         # Only types: VPN, VPN (S2S VPN) + P2S VPN
         $AddressSpaces = $IPPlan | Where-Object { $_.Type -eq "VNet" -or $_.Type -eq "VPN" -or $_.Type -eq "P2S VPN"}
         $Subnets = $IPPlan | Where-Object { $_.Type -eq "Subnet" } #VNETS split by subnets
+        $PIPs = $IPPlan | Where-Object { $_.Type -eq "PIP" } 
+        $IPPREs = $IPPlan | Where-Object { $_.Type -eq "IPPRE" }
 
         # Object reference
         # $obj = [pscustomobject]@{
@@ -5301,6 +5421,86 @@ function Export-IPPlan {
 
         # End table
         $ImagePath = Join-Path $OutputPath "icons" "snet.png"
+        $footer = "
+                </TABLE>>;
+                image = `"$ImagePath`";imagepos = `"tr`"; labelloc = `"b`";height = 2.5;];
+        }
+                "
+        $alldata = $header + $data + $footer
+        Export-AddToFile -Data $alldata
+
+        ### IPPREs
+        $header = "
+        subgraph cluster_IPPlan_IPPRE {
+            style = solid;
+            colorscheme = purples9;
+            bgcolor = 5;
+            margin = 0;
+            node [colorscheme = purples9; shape = box; color = 5; margin = 0;];
+            
+            IPPlanIPPRE [label = <
+                <TABLE border=`"0`" style=`"rounded`">
+                <TR><TD border=`"0`" align=`"left`" colspan=`"2`"><B>Public IPs in Azure</B><BR/><BR/></TD></TR>
+                <TR><TD align=`"left`"><B>Public IP</B></TD><TD align=`"left`"><B>Type</B></TD><TD align=`"left`"><B>Resource</B></TD><TD align=`"left`"><B>Resource Group</B></TD><TD align=`"left`"><B>Subscription Name</B></TD></TR>
+                <HR/>"
+        
+        # Individual Routes        
+        $data = ""
+
+        #Sort routes for easier reading
+        $IPPREs = $IPPREs | Sort-Object { [regex]::Replace($_.AddressSpace, '\d+', { $args[0].Value.PadLeft(100) }) }
+        ForEach ($IPPRE in $IPPREs ) {
+            $address = SanitizeString $IPPRE.addressSpace
+            $type = "Public IP Prefix"
+            $resource = SanitizeString $IPPRE.resource
+            $rg = SanitizeString $IPPRE.ResourceGroup
+            $subname = SanitizeString $IPPRE.SubscriptionName
+            $prefixName = SanitizeString $IPPRE.PrefixName
+            $data = $data + "<TR><TD align=`"left`">$address</TD><TD align=`"left`">$type</TD><TD align=`"left`">$resource</TD><TD align=`"left`">$rg</TD><TD align=`"left`">$subname</TD></TR>"
+        }
+
+        # End table
+        $ImagePath = Join-Path $OutputPath "icons" "ippre.png"
+        $footer = "
+                </TABLE>>;
+                image = `"$ImagePath`";imagepos = `"tr`"; labelloc = `"b`";height = 2.5;];
+        }
+                "
+        $alldata = $header + $data + $footer
+        Export-AddToFile -Data $alldata
+
+        ### PIPs
+        $header = "
+        subgraph cluster_IPPlan_PIP {
+            style = solid;
+            colorscheme = purples9;
+            bgcolor = 5;
+            margin = 0;
+            node [colorscheme = purples9; shape = box; color = 5; margin = 0;];
+            
+            IPPlanPIP [label = <
+                <TABLE border=`"0`" style=`"rounded`">
+                <TR><TD border=`"0`" align=`"left`" colspan=`"2`"><B>Public IPs in Azure</B><BR/><BR/></TD></TR>
+                <TR><TD align=`"left`"><B>Public IP</B></TD><TD align=`"left`"><B>Type</B></TD><TD align=`"left`"><B>Prefix</B></TD><TD align=`"left`"><B>Resource</B></TD><TD align=`"left`"><B>Resource Group</B></TD><TD align=`"left`"><B>Subscription Name</B></TD></TR>
+                <HR/>"
+        
+        # Individual Routes        
+        $data = ""
+
+        #Sort routes for easier reading
+        $PIPs = $PIPs | Sort-Object { [regex]::Replace($_.AddressSpace, '\d+', { $args[0].Value.PadLeft(100) }) }
+        ForEach ($PIP in $PIPs ) {
+            $address = SanitizeString $PIP.addressSpace
+            $type = "Public IP"
+            $resource = SanitizeString $PIP.resource
+            $rg = SanitizeString $PIP.ResourceGroup
+            $subname = SanitizeString $PIP.SubscriptionName
+            $prefixName = SanitizeString $PIP.PrefixName
+            $data = $data + "<TR><TD align=`"left`">$address</TD><TD align=`"left`">$type</TD><TD align=`"left`">$prefixName</TD><TD align=`"left`">$resource</TD><TD align=`"left`">$rg</TD><TD align=`"left`">$subname</TD></TR>"
+        }
+
+        # End table
+        $ImagePath = Join-Path $OutputPath "icons" "pip.png"
         $footer = "
                 </TABLE>>;
                 image = `"$ImagePath`";imagepos = `"tr`"; labelloc = `"b`";height = 2.5;];
@@ -5872,6 +6072,7 @@ function Confirm-Prerequisites {
         "keyvault.png",
         "lb.png",
         "licenses.png",
+        "ippre.png",
         #"lgw.png",
         "managed-identity.png",
         "mgmtgroup.png"
@@ -5881,6 +6082,7 @@ function Confirm-Prerequisites {
         "nic.png",
         "nsg.png",
         "peerings.png",
+        "pip.png",
         "postgresql.png",
         "private-endpoint.png",
         #"privatednszone.png",
@@ -6440,6 +6642,20 @@ function Get-AzNetworkDiagram {
                     $standardObject = $_
                     $gwObject = Get-AzVirtualNetworkGateway -ResourceGroupName $standardObject.ResourceGroupName -Name $standardObject.Name 
                     Export-GW2IPPlan -VNGW $gwObject
+                }
+
+                # Public IP
+                $PIPs = Get-AzPublicIpAddress
+                $PIPs | ForEach-Object {
+                    $PIP = $_
+                    Export-PIP2IPPlan -PIP $PIP
+                }
+
+                # Public IP Prefixes
+                $IPPREs = Get-AzPublicIpPrefix
+                $IPPREs | ForEach-Object {
+                    $IPPRE = $_
+                    Export-IPPRE2IPPlan -IPPRE $IPPRE
                 }
             }
             # Output to console